web: Extract repeated code to a new DeviceMenu component
